CAMEL-10792: Deprecate JndiRegistry in favour of JndiBeanRepository.
